Cleanup related to audio prompts and process flow
Some audio prompts need to run asynchronously (in the background) with the calling program. This is because the user will often want to press buttons without the prompt finishing. I made any prompt that expects input from the user run in subprocesses to get this effect. But other audio prompts, typlcally those that just give info or status, need to run synchronously with the caller. There are several of these that are executed usong os.system calls.

Also, there was some misc process flow related cleanup that I noticed was needed while I was testing the audio prompts. Typically this involved turning leds on or off.